api/youtube_uploader.py
import os
from google_auth_oauthlib.flow import InstalledAppFlow
from googleapiclient.discovery import build
from googleapiclient.http import MediaFileUpload
from google.oauth2.credentials import Credentials
from google.auth.transport.requests import Request

SCOPES = ["https://www.googleapis.com/auth/youtube.upload"]


def get_youtube_service(client_secrets_file: str, token_file: str):
    creds = None
    if os.path.exists(token_file):
        creds = Credentials.from_authorized_user_file(token_file, SCOPES)

    if creds and creds.expired and creds.refresh_token:
        creds.refresh(Request())

    if not creds or not creds.valid:
        flow = InstalledAppFlow.from_client_secrets_file(client_secrets_file, SCOPES)
        creds = flow.run_local_server(port=0)
        with open(token_file, "w", encoding="utf-8") as f:
            f.write(creds.to_json())

    return build("youtube", "v3", credentials=creds)


def upload_video(
    file_path: str,
    title: str,
    description: str = "",
    tags=None,
    category_id: str = "22",
    privacy_status: str = "private",
    client_secrets_file: str = "client_secret.json",
    token_file: str = "youtube_token.json",
    dimensions=(1920, 1080)
):
    service = get_youtube_service(client_secrets_file, token_file)

    # Calculate aspect ratio
    aspect_ratio = dimensions[0] / dimensions[1]
    if aspect_ratio == 9 / 16:
        privacy_status = 'public'  # Set to public for Shorts
    elif aspect_ratio == 16 / 9:
        privacy_status = 'private'  # Regular videos are private by default

    body = {
        "snippet": {
            "title": title,
            "description": description,
            "tags": tags or [],
            "categoryId": category_id,
        },
        "status": {
            "privacyStatus": privacy_status,
        },
    }

    media = MediaFileUpload(file_path, chunksize=-1, resumable=True)
    request = service.videos().insert(part='snippet,status', body=body, media_body=media)

    response = None
    while response is None:
        _, response = request.next_chunk()

    video_id = response.get("id")
    if not video_id:
        raise RuntimeError("YouTube upload completed but no video id returned")
    
    return f"https://www.youtube.com/watch?v={video_id}", video_id